Refactor product and customer models to use double for price and quantity; add event configuration for price multipliers

+++ TheMerchant/Cliente.cs
@@ -2,6 +2,6 @@ public class Cliente
 {
     public bool Sesso { get; set; }
     public TipoClasse ClasseSociale { get; set; }
-    public float Pazienza { get; set; }
-    public (Prodotto, int)? ProdottoDesiderato { get; set; }
+    public double Pazienza { get; set; }
+    public (Prodotto, double)? ProdottoDesiderato { get; set; }
 }
